backend/auth-service/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
bcrypt==4.3.0
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
Django==5.2.5
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.13.0
httpx==0.24.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pysnmp==7.1.21
pytest==7.4.2
pytest-asyncio==0.22.0
pytest-cov==4.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-nmap==0.7.1
reportlab==4.4.3
rsa==4.9.1
redis==4.5.4
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0